Code cleaning and refactoring, main() passes httpie tests
